Further refactoring of @babel/eslint-parser (#11071)
