--- eslint/babel-eslint-parser/src/babylon-to-espree/convertTokens.js
@@ -1,8 +1,0 @@
-import convertTemplateType from "./convertTemplateType";
-import convertToken from "./convertToken";
-
-export default function(tokens, code) {
-  return convertTemplateType(tokens)
-    .filter(t => t.type !== "CommentLine" && t.type !== "CommentBlock")
-    .map(t => convertToken(t, code));
-}
