+++ eslint/babel-eslint-parser/src/parse-with-scope.js
@@ -2,7 +2,7 @@ import visitorKeys from "./visitor-keys";
 import analyzeScope from "./analyze-scope";
 import parse from "./parse";
 
-export default function(code, options) {
+export default function parseWithScope(code, options) {
   const ast = parse(code, options);
   const scopeManager = analyzeScope(ast, options);
